requirements.txt
python-dotenv
discord.py
apscheduler
pytz
